pyproject.toml
[build-system]
requires = [
  "poetry==1.0.9"
]
build-backend = "poetry.masonry.api"

[tool.poetry]
authors = [
  "Kevin Amado <kamadorueda@gmail.com>",
]
classifiers = [
  "Development Status :: 6 - Mature",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]
description = "A JSON Loader with column and line numbers."
documentation = "https://kamadorueda.github.io/metaloaders"
homepage = "https://kamadorueda.github.io/metaloaders"
license = "MIT"
name = "metaloaders"
packages = [
  { include = "metaloaders", from = "src" },
]
readme = "./README.md"
repository = "https://github.com/kamadorueda/metaloaders"
version = "1.0.0"

[tool.poetry.dependencies]
lark-parser = "*"
python = "^3.6"

[tool.poetry.dev-dependencies]
bandit = "1.6.2"
mypy = "0.782"
prospector = "1.3.0"
pdoc3 = "0.8.4"
pydeps = "1.9.4"
pytest = "5.4.3"
pytest-asyncio = "0.14.0"
pytest-cov = "2.10.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/kamadorueda/metaloaders/issues"
